added sanity check for help (-h)

+++ bulkonvert.sh
@@ -4,27 +4,36 @@
 
 # Diego Project
 
-# Sanity checks ?
+# Sanity checks
 # ...
 
-if [ $# != 1 ] ; 
+if [ $1 == "-h" ]; 
+then
+  echo ""
+  echo "you must writte ./bulkonvert.sh [path]"
+  echo "the path is the way to find the folder containing the video files"
+  echo ""
+  exit 1
+else
+  if [ $# != 1 ] ; 
   then
     echo ""
     echo "bad parameter count"
-    echo "you could write ./bulkonvert.sh with one parameter only"
+    echo "you could write ./bulkonvert.sh -h for help"
     echo ""
     exit 2
-else
+  else
 # Main code
 
-  ls -1 $1*.avi > videofiles
+    ls -1 $1*.avi > videofiles
 
-  while videofile
-  do
+    while videofile
+    do
 	lengthname=${#videofile}
 	namefile=${videofile;0;$lengthname-4}
   	mencoder $videofile -o namefile.mp4 -oac mp3lame -ovc lavc -of mp4
 
-  done < videofiles
-  rm videofiles
+    done < videofiles
+    rm videofiles
+  fi
 fi
